requirements.txt
streamlit>=1.28.1
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.2
pymongo>=4.6.0
python-dotenv>=1.0.0
